+++ src/createTests.types.ts
@@ -1,7 +1,13 @@
 import type { Fn } from '@bemedev/types';
 import type { Mock } from 'vitest';
 import type { ToError_F } from './each/error.types';
-import type { TestArgs, TestErrors } from './types';
+import type {
+  ChainedFn,
+  Identity,
+  NextFn,
+  TestArgs,
+  TestErrors,
+} from './types';
 
 type ReturnR<F extends Fn> = {
   acceptation: () => void;
@@ -9,26 +15,36 @@ type ReturnR<F extends Fn> = {
   fails: (...cases: TestErrors<F>) => () => void;
 };
 
-export type _CreateTests_F = <F extends Fn>(
+export type _CreateTests_F = <
+  F extends Fn,
+  T extends NextFn<F> = Identity<F>,
+>(
   func: F,
+  transform?: T,
   toError?: ToError_F<F>,
   name?: string,
-) => ReturnR<F>;
+) => ReturnR<ChainedFn<F, T>>;
 
 export interface CreateTests_F {
-  <F extends Fn>(func: F, toError?: ToError_F<F>): ReturnR<F>;
+  <F extends Fn, T extends NextFn<F> = Identity<F>>(
+    func: F,
+    transform?: T,
+    toError?: ToError_F<F>,
+  ): ReturnR<ChainedFn<F, T>>;
 
-  withImplementation: <F extends Fn>(
+  withImplementation: <F extends Fn, T extends NextFn<F> = Identity<F>>(
     f: F,
     params: {
       instanciation: () => Promise<F> | F;
       name: string;
     },
+    transform?: T,
     toError?: ToError_F<F>,
-  ) => ReturnR<Mock<F>>;
+  ) => ReturnR<Mock<ChainedFn<F, T>>>;
 
-  withoutImplementation: <F extends Fn>(
+  withoutImplementation: <F extends Fn, T extends NextFn<F> = Identity<F>>(
     func: F,
+    transform?: T,
     toError?: ToError_F<F>,
-  ) => ReturnR<F>;
+  ) => ReturnR<ChainedFn<F, T>>;
 }