+++ src/each/error.ts
@@ -1,16 +1,21 @@
 import { expect, test } from 'vitest';
+import { identity } from '../identity';
 import { toArrayVitest } from '../toArray';
 import type { _UseErrorEach_F, UseErrorEach_F } from './error.types';
 
 export const useErrorAsyncEachCases: _UseErrorEach_F = (
   f,
+  transform = identity as any,
   toError = () => undefined,
   ...cases
 ) => {
   test.concurrent.each(toArrayVitest.error(cases))(
     '%s',
     async (_, args, _error) => {
-      const actual = async () => f(...args);
+      const actual = async () => {
+        const out = await f(...args);
+        return transform(out);
+      };
       const check = _error === undefined;
       const error = check ? toError(...args) : _error;
 
@@ -19,6 +24,11 @@ export const useErrorAsyncEachCases: _UseErrorEach_F = (
   );
 };
 
-export const useErrorAsyncEach: UseErrorEach_F = (f, toError) => {
-  return (...cases) => useErrorAsyncEachCases(f, toError, ...cases);
+export const useErrorAsyncEach: UseErrorEach_F = (
+  f,
+  transform,
+  toError,
+) => {
+  return (...cases) =>
+    useErrorAsyncEachCases(f, transform, toError, ...cases);
 };