+++ src/createTests.cov.test.ts
@@ -2,7 +2,6 @@ import { maxLength, minLength, multiChar } from '@bemedev/basicfunc';
 import { t } from '@bemedev/types';
 import { describe } from 'vitest';
 import { createTests } from './createTests';
-import { identity } from './identity';
 
 describe('CreateTests - Coverage', () => {
   type Add_F = (numb1: number, numb2: number) => number;
@@ -45,13 +44,11 @@ describe('CreateTests - Coverage', () => {
   describe('#3 => Test errors', () => {
     describe('#1 => minLength', () => {
       const { fails, success, acceptation } =
-        createTests.withoutImplementation(
-          minLength,
-          identity,
-          (min, value) => {
+        createTests.withoutImplementation(minLength, {
+          toError: (min, value) => {
             return `"${value}" is shorter or equal than ${min}`;
           },
-        );
+        });
 
       describe('#0 => Acceptation', acceptation);
 
@@ -122,7 +119,7 @@ describe('CreateTests - Coverage', () => {
         const error = 'ARGS';
         const func = async () => await Promise.reject(error);
 
-        const { fails } = createTests(func, () => error);
+        const { fails } = createTests(func, { toError: () => error });
 
         describe(
           '#1 => fails',
@@ -160,12 +157,11 @@ describe('CreateTests - Coverage', () => {
     });
 
     describe('#4 => The function has no errors, but transform throw error', () => {
-      const { fails } = createTests(
-        () => 0,
-        () => {
+      const { fails } = createTests(() => 0, {
+        transform: () => {
           throw new Error('Cov');
         },
-      );
+      });
 
       describe(
         '#1 => Fails',